+++ package.json
@@ -8,7 +8,8 @@
     "test:watch": "jest --watchAll",
     "tsc": "tsc",
     "tsc:watch": "tsc -w --pretty",
-    "tsc:check": "tsc -w --noEmit --pretty"
+    "tsc:check": "tsc -w --noEmit --pretty",
+    "start": "ts-node ./src/index.ts"
   },
   "jest": {
     "transform": {
@@ -26,11 +27,13 @@
   "license": "ISC",
   "dependencies": {
     "@phenomnomnominal/tsquery": "^2.0.0-beta.1",
+    "@types/es6-promise": "^3.3.0",
     "@types/jest": "^23.0.0",
     "@types/node": "^10.3.0",
+    "glob-fs": "^0.1.7",
     "jest": "^23.1.0",
     "ts-jest": "^22.4.6",
-    "ts-node": "^6.1.0",
+    "ts-node": "^7.0.0",
     "typescript": "^2.9.1"
   }
 }
